Make polybar work with external monitor brightness

+++ zsh/bin/reload-polybar.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # Terminate already running bar instances
-# If all your bars have ipc enabled, you can use 
+# If all your bars have ipc enabled, you can use
 polybar-msg cmd quit
 # Otherwise you can use the nuclear option:
 # killall -q polybar
@@ -9,7 +9,11 @@ polybar-msg cmd quit
 # Launch bar1 and bar2
 echo "---" | tee -a /tmp/polybar1.log /tmp/polybar2.log
 for m in $(polybar --list-monitors | cut -d":" -f1); do
-    MONITOR=$m polybar 2>&1 | tee -a /tmp/polybar1.log & disown
+    if [ "$m" = "DP-4" ] && [ -e /sys/class/backlight/ddcci13 ]; then
+        MONITOR=$m BACKLIGHT=ddcci13 polybar 2>&1 | tee -a /tmp/polybar1.log & disown
+    else
+        MONITOR=$m polybar 2>&1 | tee -a /tmp/polybar1.log & disown
+    fi
 done
 
 echo "Bars launched..."
